New class to handle user data
